+++ CArkiCliff.h
@@ -102,10 +102,10 @@ public:
 
         // Physics Setup (Same as before, Logic works fine with duplicate verts)
         btIndexedMesh meshPart;
-        meshPart.m_numTriangles = m_indices.size() / 3;
+        meshPart.m_numTriangles = (int)m_indices.size() / 3;
         meshPart.m_triangleIndexBase = (const unsigned char*)m_indices.data();
         meshPart.m_triangleIndexStride = 3 * sizeof(short);
-        meshPart.m_numVertices = m_vertices.size();
+        meshPart.m_numVertices = (int)m_vertices.size();
         meshPart.m_vertexBase = (const unsigned char*)m_vertices.data();
         meshPart.m_vertexStride = sizeof(CliffVertex);
         meshPart.m_indexType = PHY_SHORT;
@@ -167,7 +167,7 @@ public:
     {
         m_scrollAccumulator -= speed * dt;
 
-        int numRows = m_vertices.size() / 8;
+        int numRows = (int)m_vertices.size() / 8;
         float width = 3.0f;
         float zFront = 2.0f;
         float zBack = -2.0f;
@@ -241,7 +241,7 @@ public:
 
 
         device->DrawIndexedPrimitiveUP(D3DPT_TRIANGLELIST,
-            0, m_vertices.size(), m_indices.size() / 3,
+            0, (UINT)m_vertices.size(), (UINT)m_indices.size() / 3,
             m_indices.data(), D3DFMT_INDEX16,
             m_vertices.data(), sizeof(CliffVertex));
     }
